Move Scope from pyobject to frame.
