+++ vm/src/lib.rs
@@ -37,7 +37,7 @@ pub mod error;
 pub mod eval;
 mod exceptions;
 pub mod format;
-mod frame;
+pub mod frame;
 pub mod import;
 pub mod obj;
 pub mod pyobject;
